feat: add support for screen marker
