+++ packages/apps/client/src/lib/mdParser/index.ts
@@ -4,6 +4,7 @@ import { emphasisAndStrong } from './constructs/emphasisAndStrong'
 import { escape } from './constructs/escape'
 import { paragraph } from './constructs/paragraph'
 import { reverse } from './constructs/reverse'
+import { screenMarker } from './constructs/screenMarker'
 import { underlineOrHide } from './constructs/underlineOrHide'
 import { CharHandler, CharHandlerResult, NodeConstruct, ParserState } from './parserState'
 
@@ -25,6 +26,13 @@ export class ParserStateImpl implements ParserState {
 			value: this.buffer,
 		})
 		this.buffer = ''
+	}
+	setMarker = () => {
+		if (this.nodeCursor === null) throw new Error('No node available to flush buffer.')
+
+		this.nodeCursor.children.push({
+			type: 'screenMarker',
+		})
 	}
 	pushNode = (node: ParentNodeBase) => {
 		if (this.nodeCursor === null) {
@@ -65,6 +73,7 @@ export default function createParser(): Parser {
 		reverse(),
 		underlineOrHide(),
 		colour(),
+		screenMarker(),
 	]
 
 	const charHandlers: Record<string, CharHandler[]> = {}
